# Performance Profiler: Fix color accessibility