--- a/client/performance-profiler/components/charts/history-chart.jsx
+++ b/client/performance-profiler/components/charts/history-chart.jsx
@@ -30,7 +30,7 @@ const createScales = ( data, range, margin, width, height ) => {
 
 	const colorScale = d3ScaleThreshold()
 		.domain( range )
-		.range( [ '#00BA37', '#D67709', '#D63638' ] );
+		.range( [ 'var(--studio-green-60)', 'var(--studio-orange-60)', 'var(--studio-red-60)' ] );
 
 	return { xScale, yScale, colorScale };
 };