--- a/client/performance-profiler/components/charts/history-chart.jsx
+++ b/client/performance-profiler/components/charts/history-chart.jsx
@@ -30,7 +30,7 @@ const createScales = ( data, range, margin, width, height ) => {
 
 	const colorScale = d3ScaleThreshold()
 		.domain( range )
-		.range( [ 'var(--studio-green-60)', 'var(--studio-orange-60)', 'var(--studio-red-60)' ] );
+		.range( [ 'var(--studio-green-30)', 'var(--studio-orange-40)', 'var(--studio-red-50)' ] );
 
 	return { xScale, yScale, colorScale };
 };