--- a/client/performance-profiler/components/core-web-vitals-display/style.scss
+++ b/client/performance-profiler/components/core-web-vitals-display/style.scss
@@ -73,15 +73,15 @@ $blueberry-color: #3858e9;
 	margin-bottom: 24px;
 
 	&.good {
-		color: #00ba37;
+		color: var(--studio-green-60);
 	}
 
 	&.needs-improvement {
-		color: #d67709;
+		color: var(--studio-orange-60);
 	}
 
 	&.bad {
-		color: #d63638;
+		color: var(--studio-red-60);
 	}
 }