--- a/client/performance-profiler/components/performance-score/style.scss
+++ b/client/performance-profiler/components/performance-score/style.scss
@@ -127,15 +127,15 @@
 			background: var(--studio-orange-40);
 
 			&.poor {
-				background: var(--studio-red-60);
+				background: var(--studio-red-50);
 			}
 
 			&.neutral {
-				background: var(--studio-orange-60);
+				background: var(--studio-orange-40);
 			}
 
 			&.good {
-				background: var(--studio-green-60);
+				background: var(--studio-green-30);
 			}
 		}
 	}