… client/performance-profiler/icons/thumbs.tsx