Part of #101301

## Proposed Changes

On the Performance Profiler page, increases the darkness of the green/orange/red colors to the `60` shade.

## Why are these changes being made?

This follows the new [guidance](https://github.com/Automattic/wp-calypso/issues/101301#issue-2916247365) to use the `60` shade for all standalone green color usage. Following the same reasoning, red and orange colors also need the same rule applied. Though, note that the three colors don't have the same contrast characteristics for when they're set on a `#fdfdfd` background:

|        | 50 | 60 |
|--------|----|----|
| green  | ❌  | ✅  |
| orange | ❌  | ✅  |
| red    | ✅  | ✅  |

(Of course, the `50` shades are all ✅ when set on `#ffffff`, as is the case on this Performance Profiler page.)

Similar to https://github.com/Automattic/wp-calypso/pull/101391#issuecomment-2726031195, I'll flag this as another case where there are potential downsides to uniformly increasing these colors to `60`. The difference between red and orange is less distinguishable. Let me know if we want to try something else, like mix and matching shades @Automattic/dotcom-design.

The existing styles here use `30` and `40` shades in some places, so we need to bump those all to `50`, at the very least.

### Screenshots

| Before | After |
|--------|-------|
|<img src="https://github.com/user-attachments/assets/69d29df2-6e72-4471-9d43-652e059405de" alt="Before the fold, before" width="1130">|<img src="https://github.com/user-attachments/assets/e58cf5df-3aff-4312-8e3c-83be924a0e1b" alt="Before the fold, after" width="1130">|

| Before | After |
|--------|-------|
|<img src="https://github.com/user-attachments/assets/d60ea9aa-11b6-459d-afe1-67baa6440ec7" alt="Recommendations, before" width="1092">|<img src="https://github.com/user-attachments/assets/b4ed7a22-497f-48c7-b907-8bfe1e374e3e" alt="Recommendations, after" width="1092">|

| Before | After |
|--------|-------|
|<img src="https://github.com/user-attachments/assets/0e8bb825-1646-4269-8a87-25e9d790e023" alt="AI callout actions, before" width="977">|<img src="https://github.com/user-attachments/assets/ee47c562-0767-4ef4-86ab-808e78d54dcc" alt="AI callout actions, after" width="977">|


## Testing Instructions

Go to `/speed-test-tool?url=https%3A%2F%2Fwordpress.com%2F&hash=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6NzA1N30.7KXaBL1qcm0y0GqvM8ymJ0pbST9ls6P09_hri8EaKSI`

## Pre-merge Checklist

<!--
Complete applicable items on this checklist **before** merging into trunk. Inapplicable items can be left unchecked.

Both the PR author and reviewer are responsible for ensuring the checklist is completed.
-->

- [ ] Has the general commit checklist been followed? (PCYsg-hS-p2)
- [ ] [Have you written new tests](https://wpcalypso.wordpress.com/devdocs/docs/testing/index.md) for your changes?
- [ ] Have you tested the feature in Simple (P9HQHe-k8-p2), Atomic (P9HQHe-jW-p2), and self-hosted Jetpack sites (PCYsg-g6b-p2)?
- [ ] Have you checked for TypeScript, React or other console errors?
- [ ] Have you used memoizing on expensive computations? More info in [Memoizing with create-selector](https://github.com/Automattic/wp-calypso/blob/trunk/packages/state-utils/src/create-selector/README.md) and [Using memoizing selectors](https://react-redux.js.org/api/hooks#using-memoizing-selectors) and [Our Approach to Data](https://github.com/Automattic/wp-calypso/blob/trunk/docs/our-approach-to-data.md)
- [ ] Have we added the "[Status] String Freeze" label as soon as any new strings were ready for translation (p4TIVU-5Jq-p2)?
    - [ ] For UI changes, have we tested the change in various languages (for example, ES, PT, FR, or DE)? The length of text and words vary significantly between languages. 
- [ ] For changes affecting Jetpack: Have we added the "[Status] Needs Privacy Updates" label if this pull request changes what data or activity we track or use (p4TIVU-aUh-p2)?