@crisbusquets 2025-03-24
> for when they're set on a #fdfdfd background:

@jasmussen where did we landed with the background color? #fdfdfd is _almost_ white. cc @matt-west .