@mirka 2025-03-25
Not sure what's the absolute darkest "light" background we have in production though, for example on the [Domains page](https://wordpress.com/start/domain/domain-only) it's `#fdfdfd` but on [Sites](https://wordpress.com/sites) we have a `#f6f7f7` sidebar and `#f8f8f8` hover on DataViews rows.