@jasmussen 2025-03-25
Short answer: we can go with Studio Green 60. It's what @nuriapenya suggested in a related context, it makes it connected to a variable and therefore easier to update in the future, and it will work on any of the grays that are being discussed. (I checked, that shade of green will work on a gray as dark as #dbdbdb, assuming I'm not missing nuance.)

---

To look a bit ahead, there's an open question about a singler background color to employ across onboarding and in accents of a new hosting dashboard, tracked here: https://github.com/Automattic/wp-calypso/issues/101133. For the moment, the plan (correct me if I'm wrong @matt-west) is a white background. But even without that, I expect in most contexts, especially those with green numbers, they'll be seen on cards (Card or future DataSummary components) or in DataViews, which have white backgrounds.