@nuriapenya 2025-03-25
For graphical elements a contrast of 3:1 should be enough, which means there's no need to update the bar or the icons. But, there are some instances of text that clearly need to be updated:

![CleanShot 2025-03-25 at 12 33 57@2x](https://github.com/user-attachments/assets/ca557e8c-a740-4e7d-9cf8-66fb9163eff0)
![CleanShot 2025-03-25 at 12 34 38@2x](https://github.com/user-attachments/assets/ac9c77eb-c8fb-4451-ab7b-c9d4c4513a61)

Since the purpose of this page is to compare between the three tiers (red, orange, green) perhaps it's worth doing an exception and say`studio-50` it's enough. Most of them seem to be over white background and rather big text, so going to 60 might not give enough differentiation. 

Would that be an acceptable exception? :)