@jasmussen 2025-03-25
> Would that be an acceptable exception? :)

If we're looking purely from WCAG, it's on the edge, though I'd overall lean the interpretation towards a _yes if we feel strongly about it_. For the context, the 3:1 contrast works for UI and "large text", which is suggested as 14pt (18.66px) when _bold_, or 18pt (24px) and normal weight. This is 20px non-bold so it's on the edge.

But extracting ourselves from some of those guidelines and looking more on the larger intent, there's something super easy about just embracing studio 60 green for this across the board, knowing it's always "safe". Which means, if we want to do something different, we just need to _want to do it_. And if we want that vibrancy, go for it.

Another alternative: if a 24px font size works for you for these numbers, most of this nuance disappears :)