@matt-west 2025-03-25
I agree with @nuriapenya. We should use the 60 colors for text, but the progress bars and chart icons should retain the existing colors.

The chart uses shape to differentiate the point types, so color is less relevant in that context.

> For the moment, the plan is a white background.

That's correct.